README.md
Champlain Problem Library (CPL) repository.

This is the library of WeBWorK problems created by members of the mathematics department of Champlain College Saint-Lambert.

This repository has three parts:
          1. CPL-Contrib
          2. CPL-Pending
          3. champlain-problem-library

The contents of champlain-problem-library are the problems one can browse via the Library Browser in WeBWorK by clicking on the Champlain button and browsing the subdirectories by subject-topic. Updates to the CPL will occur when classes are NOT in session. Although this whole repository is named champlainww-cpl, we will use CPL to refer to just this part.

New contributions of problems to the Champlain Problem Library (CPL) should go into  the author's subdirectory of CPL-Contrib, where the name of the subdirectory is the author's initials capitalized.  The author retains control of these files, both in style and in content.

Problems that are being tested or reviewed before inclusion in the CPL are copied temporarily into CPL-Pending.